move helper classes to Misc
