+++ src/Runtime/Optimized/Misc/IWasmSpinLock.ts
@@ -1,0 +1,6 @@
+/**
+ * Interface for a WASM spin lock
+ */
+export interface IWasmSpinLock {
+    wait(): void;
+}
