+++ src/Runtime/Optimized/Misc/IWasmTypedArray.ts
