+++ src/Runtime/Optimized/Misc/wasmSharedTypedArray.ts
