+++ src/Runtime/Optimized/Misc/wasmSpinlock.ts
@@ -1,3 +1,4 @@
+import type { IWasmSpinLock } from "./IWasmSpinLock";
 import type { IWasmTypedArray } from "./IWasmTypedArray";
 
 // const enum SpinLockConstants {
@@ -7,7 +8,7 @@ import type { IWasmTypedArray } from "./IWasmTypedArray";
 /**
  * Spinlock for WASM runtime synchronization
  */
-export class WasmSpinlock {
+export class WasmSpinlock implements IWasmSpinLock {
     private readonly _lock: IWasmTypedArray<Uint8Array>;
 
     /**
