+++ src/Runtime/Optimized/Misc/wasmTypedArray.ts
