+++ src/Runtime/Optimized/mmdWasmRuntimeBone.ts
@@ -8,7 +8,7 @@ import type { IMmdRuntimeBone } from "../IMmdRuntimeBone";
 import type { IMmdRuntimeLinkedBone } from "../IMmdRuntimeLinkedBone";
 import type { MmdWasmInstance } from "./mmdWasmInstance";
 import type { MmdWasmRuntime } from "./mmdWasmRuntime";
-import { type WasmBufferedArray, WasmBufferedArraySpan } from "./wasmBufferedArray";
+import { type WasmBufferedArray, WasmBufferedArraySpan } from "./Misc/wasmBufferedArray";
 
 /**
  * Bone for MMD WASM runtime
